Added comments to the newly created classes, and moved record test class to appropriate package
